refactor(lib): rename .mjs files to .ts
