+++ src/lib/vector-search.ts
